flux subroutine was simplified

+++ 3D_solver/TBL/mod_globals.f90
@@ -49,7 +49,7 @@ module mod_globals
   real(2), parameter         :: id_scheme   = 0
   integer, parameter         :: id_sensor   = 1
   real(8), parameter         :: threshold   = 0.4d0
-  integer(kind=8), parameter :: id_accuracy = 0
+  integer(kind=4), parameter :: id_accuracy = 0
   integer(kind=2), parameter :: id_tvd      = 0
   integer(kind=2), parameter :: id_keep     = 0
   integer(kind=4), parameter :: id_slau     = 0
